refactor(core,api-v3): seperate structs and enums from `NativeMemory`
