+++ source/core/Enums/CScriptResourceTypeNameIndex.cs
@@ -1,0 +1,9 @@
+namespace SHVDN
+{
+    internal enum CScriptResourceTypeNameIndex : ushort
+    {
+        Checkpoint = 6,
+        RelGroup = 20,
+        ScaleformMovie = 21,
+    }
+}
