+++ source/core/Enums/EntityTypeInternal.cs
@@ -1,0 +1,12 @@
+namespace SHVDN
+{
+    internal enum EntityTypeInternal
+    {
+        Invalid = 0,
+        Building = 1,
+        AnimatedBuilding = 2,
+        Vehicle = 3,
+        Ped = 4,
+        Object = 5
+    }
+}
