+++ source/core/Enums/ModelInfoClassType.cs
@@ -1,0 +1,13 @@
+namespace SHVDN
+{
+    internal enum ModelInfoClassType
+    {
+        Invalid = 0,
+        Object = 1,
+        Mlo = 2,
+        Time = 3,
+        Weapon = 4,
+        Vehicle = 5,
+        Ped = 6
+    }
+}
