+++ source/core/Enums/PoolType.cs
@@ -1,0 +1,9 @@
+namespace SHVDN
+{
+    internal enum PoolType
+    {
+        Generic,
+        Vehicle,
+        Projectile,
+    }
+}
