+++ source/core/Enums/Vehicle/VehicleStructClassType.cs
@@ -1,0 +1,22 @@
+namespace SHVDN
+{
+    internal enum VehicleStructClassType
+    {
+        None = -1,
+        Automobile = 0x0,
+        Plane = 0x1,
+        Trailer = 0x2,
+        QuadBike = 0x3,
+        SubmarineCar = 0x5,
+        AmphibiousAutomobile = 0x6,
+        AmphibiousQuadBike = 0x7,
+        Heli = 0x8,
+        Blimp = 0x9,
+        Autogyro = 0xA,
+        Bike = 0xB,
+        Bicycle = 0xC,
+        Boat = 0xD,
+        Train = 0xE,
+        Submarine = 0xF
+    }
+}
