+++ source/core/Structs/DragDrawable.cs
@@ -1,0 +1,11 @@
+using System.Runtime.InteropServices;
+
+namespace SHVDN
+{
+    [StructLayout(LayoutKind.Explicit)]
+    internal unsafe struct FragDrawable
+    {
+        [FieldOffset(0x18)]
+        internal CrSkeletonData* crSkeletonData;
+    }
+}
