+++ source/core/Structs/FwScriptGuidPool.cs
@@ -1,0 +1,31 @@
+using System.Runtime.InteropServices;
+using System.Runtime.CompilerServices;
+
+namespace SHVDN
+{
+    // Note: actually this struct is supposed to point the same struct type as `FwBasePool` in this source code
+    // file, but needs to be careful when refactoring.
+    [StructLayout(LayoutKind.Explicit)]
+    internal struct FwScriptGuidPool
+    {
+        // The max count value should be at least 3072 as long as ScriptHookV is installed.
+        // Without ScriptHookV, the default value is hardcoded and may be different between different game versions (the value is 300 in b372 and 700 in b2824).
+        // The default value (when running without ScriptHookV) can be found by searching the dumped exe or the game memory with "D7 A8 11 73" (0x7311A8D7).
+        [FieldOffset(0x10)]
+        internal uint maxCount;
+        [FieldOffset(0x14)]
+        internal int itemSize;
+        [FieldOffset(0x18)]
+        internal int firstEmptySlot;
+        [FieldOffset(0x1C)]
+        internal int emptySlots;
+        [FieldOffset(0x20)]
+        internal uint itemCount;
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        internal bool IsFull()
+        {
+            return maxCount - (itemCount & 0x3FFFFFFF) <= 256;
+        }
+    }
+}
