+++ source/core/Structs/PedPersonality.cs
@@ -1,0 +1,15 @@
+using System.Runtime.InteropServices;
+
+namespace SHVDN
+{
+    [StructLayout(LayoutKind.Explicit, Size = 0xB8)]
+    internal struct PedPersonality
+    {
+        [FieldOffset(0x7C)]
+        internal bool isMale;
+        [FieldOffset(0x7D)]
+        internal bool isHuman;
+        [FieldOffset(0x7F)]
+        internal bool isGang;
+    }
+}
