+++ source/core/Structs/ScreenInfo.cs
@@ -1,0 +1,18 @@
+using System.Runtime.InteropServices;
+
+namespace SHVDN
+{
+    [StructLayout(LayoutKind.Explicit, Size = 0x30)]
+    internal struct ScreenInfo
+    {
+        // these fields should be in pixel coordinates
+        [FieldOffset(0x14)]
+        internal uint left;
+        [FieldOffset(0x18)]
+        internal uint right;
+        [FieldOffset(0x1C)]
+        internal uint top;
+        [FieldOffset(0x20)]
+        internal uint bottom;
+    }
+}
