+++ source/scripting_v3/GTA/Weapons/WeaponCollection.cs
@@ -70,7 +70,7 @@ namespace GTA
 
                 unsafe
                 {
-                    return ((SHVDN.NativeMemory.RageAtArrayPtr*)(pedInventoryAddr + 0x18))->size;
+                    return ((SHVDN.RageAtArrayPtr*)(pedInventoryAddr + 0x18))->size;
                 }
             }
         }
@@ -453,14 +453,14 @@ namespace GTA
                         return null;
                     }
 
-                    var weaponInventoryArray = (SHVDN.NativeMemory.RageAtArrayPtr*)(pedInventoryAddr + 0x18);
+                    var weaponInventoryArray = (SHVDN.RageAtArrayPtr*)(pedInventoryAddr + 0x18);
                     if (index >= weaponInventoryArray->size)
                     {
                         return null;
                     }
 
                     ulong itemAddress = weaponInventoryArray->GetElementAddress(index);
-                    SHVDN.NativeMemory.ItemInfo* weaponInfo = *(SHVDN.NativeMemory.ItemInfo**)(itemAddress + 0x8);
+                    SHVDN.ItemInfo* weaponInfo = *(SHVDN.ItemInfo**)(itemAddress + 0x8);
                     if (weaponInfo == null)
                     {
                         return null;
